requirements.txt
matplotlib
streamlit
requests
beautifulsoup4
plotly
pandas
fpdf
altair
watchdog
psutil
numpy
scipy
pytest
Pillow
biopython
lxml
git+https://github.com/pre-commit/pre-commit-hooks.git
py3Dmol
dssp
freesasa-python
molprobity
prosa
qmean
voromqa
pdbfixer
modeller
fastapi
uvicorn
flask
